rename languages module to lang to match spacy
