+++ tokenizer/lib/spacy/lang/la/normalizations.py
