+++ tokenizer/lib/spacy/lang/mapper.py
@@ -17,7 +17,7 @@ LANGUAGE = {
     'fi': { "name": "Finnish" },
     'fr': { "name": "French" },
     'ga': { "name": "Irish" },
-    'grc': { "path": "tokenizer.lib.spacy.languages.grc", "name": "AncientGreek"},
+    'grc': { "path": "tokenizer.lib.spacy.lang.grc", "name": "AncientGreek"},
     'gu': { "name": "Gujarati" },
     'he': { "name": "Hebrew" },
     'hi': { "name": "Hindi" },
@@ -30,7 +30,7 @@ LANGUAGE = {
     'ja': { "name": "Japanese" },
     'kn': { "name": "Kannada" },
     'ko': { "name": "Korean" },
-    'la': { "path": "tokenizer.lib.spacy.languages.la", "name": "Latin"},
+    'la': { "path": "tokenizer.lib.spacy.lang.la", "name": "Latin"},
     'lb': { "name": "Luxembourgish" },
     'lij': { "name": "Ligurian" },
     'lt': { "name": "Lithuanian" },
